SuperApi/requirements.txt
django~=2.2
appdirs==1.4.4
asgiref==3.4.1
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
colorama==0.4.4
distlib==0.3.1
Django==3.2.13
django-cors-headers==3.11.0
djangorestframework==3.13.1
filelock==3.0.12
importlib-metadata==4.0.1
iniconfig==1.1.1
mysql-connector-python==8.0.26
packaging==20.9
Pillow==9.1.0
pip==21.1
pipenv==2020.11.15
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
pytz==2021.3
setuptools==47.1.0
six==1.15.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.7.4.3
virtualenv==20.4.4
virtualenv-clone==0.5.4
whitenoise==6.0.0
zipp==3.4.1
